// src/packages/bui-react/src/components/AvatarBlock/AvatarBlock.types.ts
import React from "react";
import type * as TAvatar from "components/Avatar/Avatar.types";
import type * as TStack from "components/Stack/Stack.types";
import type * as G from "types/global";
import type { Responsive, Mixin } from "@bookingcom/bui-core/types";

export type Size = "small" | "medium" | "large";

export type Props = {
  size?: Responsive<Size>;
  color?: "inherit";
  title: React.ReactNode;
  subtitle?: React.ReactNode;
  avatar: Omit<TAvatar.Props, "accessibilityLabel" | "size">;
  className?: string;
  attributes?: G.Attributes<"div">;
  mixin?: Mixin<{ height: false }, TStack.Props["mixin"]>;
};